Cleanup and test AnnotatedPatchedFile
The AnnotatedPatchedFile was and is called PatchFile in the original
'annotate.py'.  Currently only the constructor is being tested; the
process() method is a simple stub.

NOTE that to make it easier to do partial cleanup, the cleaned up code
with docstrings and typing information is being added, step by step, to
a temporary file 'new_annotate.py'.  It would replace 'annotate.py' when
ready.

Tests for cleaned up code are present in 'test_annotate.py'.  Testing is
done by the way of annotate_single_diff() method, copied from
'single_diff_annotator.py' file in the 'main' branch.