Add and test the Bug class
The Bug class, representing a singlle bug in a dataset, was added to
'new_annotate.py' from 'annotate.py'.

The code was modernized to use pathlib instead of os.path etc.  The
interface was cleaned up, with the constructor now doing the reading,
parsing, and annotating, while saving is done with the save() method.
The constructor does not create any files, now.

The class was made a bit more configurable, both in regard to expected
directory structure, and in choosing where to save the results.  It
should be easier to use.

Added test for Bug constructor, using patch file created for other test.
TODO: Add test for the save() method.