Make new_annotate.py a Python script
This copies the code from the original 'annotate.py' that makes it
a Python script, and adds support for command line arguments.
Uses she-bang to mark it as script.

The code had to be adjusted for changes in the behavior of the Bug
class.  It was manually checked that the script works as it should.

TODO?: Move from 'click' to 'typer': https://typer.tiangolo.com/