Add test_cli.py
Adds test for annotate.py script via typer.testing.CliRunner (the script
uses 'typer' module for handling command line arguments).
https://typer.tiangolo.com/tutorial/testing/

Note that runner.invoke must be passed list of _str_ as second argument
(same as if it were invoked from the command line), using Path results
in error.

Only test for the "patch" subcommand is added, as "dataset" subcommand
currently does not allow saving the output files in a temporary
directory that is created for tests.

For the test to pass, the patch() function must ensure that the
directory where the JSON file with annotation data will be saved exists,
creating it if necessary.