Add and test line_callback
Currently line_callback is added as a static (global) field in
AnnotatedPatchedFile, though there might be a better place for it.

The result of line_callback takes precedence; if you want to leave the
detection to built-in code, the callback should return None.

Test that very simple callback and also a more complex callback work
without problems.