Add and test make_line_callback() method
The make_line_callback() is currently added as a static method to
AnnotatedPatchedFile class.  It can be used to generate line callback
code from string.  It's intended use is to help support --line-callback
command line options.

This method is based on the local make_callback() function inside the
_handle_arg_callback() method of RepoFilter class in git-filter-repo:
https://github.com/newren/git-filter-repo/blob/main/git-filter-repo#L2839

The test uses multi-line body of callback function, and was added not as
separate test, but as part of test_line_callback_whitespace() test in
test_annotate.py test file.

While at it, remove failed attempt at creating closure-like callback.
It is not necessary for requested functionality.