Add and test --line-callback global option
Note that you can pass as a callback either a body of a function as
string (possibly mmultiline string), or as path to file with body of
said function.

The parse_line_callback() function used to parse --line-callback option
argument is based on the local handle(type) function inside the
_handle_arg_callback() method in RepoFilter class in git-filter-repo:
https://github.com/newren/git-filter-repo/blob/main/git-filter-repo#L2843

TODO: support writing whole function (don't add "header" then).