The --line-callback can be given full function
Instead of passing just the bodu of the function, for example:

   return None

one can now pass the code that defines LineCallback-like function,
for example

   def nop_callback(tokens):
       return None

This is especially useful if the callback function is loaded from
a file.

NOTE however that the check for the second form uses regexps, and
requires function definition to be the first line in the file
(be in the first line of the source code for line callback function).

Adds simple test for the new case.