Make PURPOSE_TO_ANNOTATION a global variable
PURPOSE_TO_ANNOTATION was formerly a static field in the AnnotatedHunk
class.  Having it to be a global variable makes it much easier to find;
it is simply an easier API, as you don't need to remember which of the
different classes contains it.

While at it, treat '' and "" values as the empty value is treated.
This is to help avoid mistakes in command line handling.