Improve debuggability of annotate_single_diff()
Split single try: ... except: into two to better find where was the
problem in a more precise way, that is in which part of the process
there was a problem.

Because processing the patch (adding annotations) has more places
where there can be an error, than using unidiff for parsing the patch,
therefore annotate.py would now print a traceback.

While at it, rename local variable to aboid shadowning the name 'patch'
from outer scope.

TODO: even more detailed check: in which patch file there was an error.