Add and test --ext-to-language global option
Note that there is a bit of code duplication between the code for
handling --purpose-to-annotation and --ext-to-language, but they are
different enough that it would be difficult to extract common code.

The test is a very basic one, chacking that the 'annotate.py' script
runs with --ext-to-language option + checkiung the output, but it does
not check if it works as intended - that is what API tests in
test_annotate.py are for.