pyproject.toml: Add [project.scripts] section
Currently it contains only two of the three planned scripts.  The names
of scripts are not final, and might change in the future.

Based on the "Building a Package" section in 'typer' documentation

  https://typer.tiangolo.com/tutorial/package/

but without using Poetry as the build tool, and instead taking
the inspiration from the "Creating executable scripts" section in
the Python Packaging User Guide:

  https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#creating-executable-scripts