generate_patches.py: Add clone_repository to GitRepo
The clone_repository() static method was copied from the
https://github.com/ncusi/MSR_Challenge_2024/blob/main/src/utils/git.py
with the main change of moving typing information from the docstring
to parameters.  While at it, make `dissasociate` parameter `bool` type,
instead of `bool | None` aka. `Optional[bool]`.

This method would be useful for testing.  We cannot assume that test
would be run from the python-diff-annotator repository.