Add test_generate_patches.py with test for clone_repository()
Running the test have shown that on MS Windows there might occur a
strange error ("getadrinfo() thread failed to start"), which might be
caused by Windows Defender, or might be a transient error - but it
occurs only when running `git clone` process with specified custom
environment variables.  Rerun cloning without `env` in such case.

Note that PyCharms linter shows the following warning for the
`working_dir` parameter of GitRepo.clone_repository() method:

  Expected type 'str | None' (matched generic type 'PathLike ?: str | bytes | Path | PathLike | None'), got 'Path' instead

which I think is incorrect, but needs checking (for example with mypy,
or with Pyre, or Pytype).