generate_patches.py: Fix signature of format_patch() method
The *args can be used to pass variable number of _positional_ arguments
to the function, but it would require to have *args before any keyword
arguments, and require keyword arguments to not be positional.  We might
go back to the solution just mentioned, but for now replace *args with
revision_range parameter that accepts either iterable (list or tuple),
or a single string.

If passed a single string, it behaves like <since> argument to
`git format-patch`; if passed iterable, it behaves like
<revision-range>.  If passed nothing for revision_range, format_patch()
would create single patch from the most recent commit.

While at it, add 'encoding' parameter to subprocess.run() invocation.