generate_patches.py: Add unidiff() to GitRepo
This method is copied from utils.git module in ncusi/MSR_Challenge_2024
https://github.com/ncusi/MSR_Challenge_2024/blob/main/src/utils/git.py

It is a start of series of revisions that would introduce API for
generating the diff annotation processing pipeline.  The GitRepo class
will be able to generate list of unidiff.PatchSets, and pass it to
annotation process, instead of having to save patches to disk.