generate_patches.py: Add and test log_p() method of GitRepo
The log_p() method is a generator function, in order to make it possible
to get partial results as fast as possible, and not to wait for the
whole `git log` process to finish walking the revisions.

It reads from `git log -p -z` line by line; it makes it quite eazy to
find the end of one commit and beginning of the other.  Possibly
a better solution would be to use non-blocking readinto1 into bytes-like
buffer.

StringIO is used as kind of a string builder instead of repeated
concatenation with +=, or gathering data into list and then using
join().  This trick was taken from the following article:
https://realpython.com/python-string-concatenation/#concatenating-strings-with-stringio

.
Basic test for log_p() was added to test_generate_patches.py.
While at it, the exploratory test added in previous commit was marked
to be skipped.