generate_patches.py: Modify log_p() to provide commit_id, and fix it
If the log_p() method is invoked with wrap=True, the unidiff.PatchSet
is now extended with the 'commit_id' attribute that contains commit id
(a SHA-1 in hex) extracted from commit data.  Test that it works
correctly in test_generate_patches.py.

While at it, *fix* a serious bug, namely that in wrap=True mode returned
PatchSets that were empty, which was because the commit_data variable
(of StringIO type) was not rewound to the beginning before passing it
to PatchSet constructor.

NOTE: in the future the commit message might be parsed, and returned as
<parsed commit>, <parsed patch/changeset> tuple, named tuple, typed
dict, or data class -- if requested.