TODO.md: Update information about calling Ruby from Python
The GitHub Linguist library is written in Ruby; the command line
application it provides, namely 'github-linguist', does not provide
finding language based on filename alone.

RbCall / rb_call is without license, and uses RPC client and server
to translate between Python and Ruby (using MessagePack for
serialization).

Rython installs cleanly, but import fails with SyntaxError; because last
version of this package is from 2009, it might be Python 2.x only:

  In [1]: import rython
  Traceback (most recent call last):

    File /usr/local/lib/python3.11/dist-packages/IPython/core/interactiveshell.py:3508 in run_code
      exec(code_obj, self.user_global_ns, self.user_ns)

    Cell In[1], line 1
      import rython

    File [...]/.venv/lib/python3.11/site-packages/rython/__init__.py:191
      except socket.error, e:
           ^
  SyntaxError: multiple exception types must be parenthesized