annotate.py: Add --update-languages option, on by default
Because douban/linguist project (https://pypi.org/project/linguist/)
and its retanoj/linguist fork (https://github.com/retanoj/linguist)
do not keep up to date with 'languages.yml' from the GitHub Linguist
project (https://github.com/github-linguist/linguist), with latest
update of the 'languages.yml' file from 2020 (or earlier), the
annotate.py script now by default uses it's own copy of the file.
This behavior can be turned off with --no-update-languages option.

To compare old version and new version of 'languages.yml' in the command
output the size is used; the modification time would tell us when the
package was installed, not when the file was last updated.

.
Additionally, move ensuring that difannotator.languages.Language.yaml
is absolute path from the _read() to the __init__() method.  Not that
it changes behavior, because _read() is invoked by __init__().

While at it, remove comment with information that was duplicated
in the output message, and add comments configuring PyCharms linter.