languages.py: Introduce and use PATTERN_TO_PURPOSE mapping
The new PATTERN_TO_PURPOSE mapping replaces PROJECT_MANAGEMENT and
DOC_PATTERNS lists (which should have been sets, instead).  This change
would make it easier to introduce command line option to configure this
behavior.

While at it, add a few more filenames (and patterns) of project files,
and for each project file add a comment with explanation what build or
package manager uses that file, and/or with what programming language it
is associated with.

Sources to consider:
- https://libraries.io/#search-by-language
- https://en.wikipedia.org/wiki/List_of_software_package_management_systems#Application-level_package_managers