languages.py: Replace TEXT_FILES with FILENAME_TO_LANGUAGES
The FILENAME_TO_LANGUAGES functions in the same way as reverse mapping
from filename (basename) to a list of languages extracted from
'languages.yml'.  It also noow includes a few special cases that were
previously hardcoded as conditionals in _path2lang().

This change, like the one in the previous commit, would make it easier
to introduce command line option to configure this behavior.