annotate.py: Move sanitization from callback to common()
Extract the code that ensures that all keys in the EXT_TO_LANGUAGES dict
(which represent extensions) beging with a dot (.) from the
to_language_mapping_callback() function to the common() function that
processes arguments common to all (sub)commands.

This is done in preparation for adding --filename-to-language, in order
to be able to use to_language_mapping_callback() as a common code
between it and --ext-to-language.  Additionally, the new placement
ensures that if there were any mistakes in EXT_TO_LANGUAGES definition,
they would be corrected.

Also, delete old entries with key / extension without the leading dot
(.).

While at it, choose better names for loop variables in common(), and
separate blocks of code with empty lines for better readibility.