annotate.py: Extract to_simple_mapping_callback()
This function is a more generic version of what was originally the
purpose_to_annotation_callback() function.  The second function now
silmply calls to_simple_mapping_callback() with appropriate arguments,
and is still used to implement parsing values for --purpose-to-annotation
option.

The to_simple_mapping_callback() was made more generic to be able to
implement the callback for planned --pattern-to-purpose option.