annotate.py: Add `fan_out` parameter to `Bug.from_dataset()`
Like the previous commit, this feature would be useful in implementing
support for the '--use-fanout' parameter, but this time for a different
subcommand, namely 'diff-annotate dataset'.

The file used for the new test is a test file from a different test,
just in different directory, and with different pathname and name.