annotate.py: Add {annotations,patches}_dir to BugDataset
All constructors of the BugDataset class now take optional 'patches_dir'
and 'annotations_dir', to remember them, and to pass them to individual
Bugs.

This made it possible to easy test that 'fan_out' parameter, added to
BugDataset constructors in the previous commit, works as intended.

While at it, add a "TODO?" comment to BugDataset class.