annotate.py: Add '--uses-fanout' option for 'dataset' subcommand
This option allows to use dataset generated by the 'diff-generate'
script (which is created out of 'generate_patches.py') with the
'--use-fanout' option.

To make it easier to test (in test_cli.py), add also '--patches-dir'
and '--annotations-dir' options to be able to configure the exact
structure of paths in the dataset.

While at it, improve formatting of dataset() signature / declaration.