languages.py: "Fix" the test_languages.py test
When running the whole test suite (with `pytest`), the test_languages.py
test fails... because languages.EXT_TO_LANGUAGES is empty.  When running
only the failing test (with `pytest -s ./tests/test_languages.py`), the
test works.

This should not happen.  For now, skip the test if it detect this
strange situation.

Problem was analyzed with the help of some debugging print()s, now
commented out.

.
While at it, rename `languages` variable in test_Languages() function in
tests/test_languages.py to `langs`, to avoid it shadowing the
`languages` module / package.