notebooks: Add 01-gather_data-explore.ipynb notebook
Currently it uses external data file that is not in the repository,
and is not available online (for example with DVC or Git-LFS).
This data file contains information about file purposes.

Add packages that the new notebook uses to notebooks/requirements.txt

NOTE: The notebook is added without stripping outputs.