annotate.py: Store pre-/post-images in AnnotatedPatchedFile
This will allow in the future to be able to have an option to parse
whole files, and not just file fragments (change + context lines).
It may happen that opening of string or comment is outside context
lines, and paring only a fragment of code would lead to wrong lex
result.

The result of lexing is cached, to avoid costly recomputation (per
hunk, or when file is copied into multiple target files).

Also adds some basics tests.