annotate.py: Add hunk_tokens_for_type() method to AnnotatedPatchedFile
The hunk_tokens_for_type(), which gives results of lexing pre-image
or post-image contents of the file, for each file in chunk: either
removed (lexing pre-image) or added (lexing post-image).

Change the format of storing the result of lexing to match what is used
in the body of AnnotatedHunk::process() method.  This way result of
lexing pre- or post-image (respectively) would be able to be used in
place of lexing fragment of pre- or post-image recovered from the diff
(i.e. changed lines plus lines of context).

Add some tests that show that we get the same line in process() and in
hunk_tokens_for_type(), even if the result of the parse is different.

NOTE: First use of parametrized test.  Lots of commented out debug
prints in the added checks.