utils.git: Add typing info to signatures, where was missing
This makes the style more uniform, as previously some methods had typing
information in the signature, some in doctring (this was not removed),
and some were missing typing information altogether.

Also, switch to the syntax compatibile with Python 3.7 (minimum
supported Python version, according to pyproject.toml), which means
replacing:
 - a | b   ->  Union[a, b]   (union type expressions require 3.10)
 - list[a] ->  List[a]       (this and similar types require 3.9)

While at it, modify the example in file docstring.