annotate.py: Fix hunk processing with sources, fix tests
The process() method in AnnotatedHunk now expects keys in the
`tokens_group` variable to be index into lines in hunk.  This required
adjusting the hunk_tokens_for_type() method in AnnotatedPatchedFile
to use 0-based index into hunk line as keys, instead of using 1-based
line number in pre-/post-image.

The case for where there is no pre-/post-image had to also be adjusted
to map from line number is composed source to line number in whole hunk
in `tokens_group` variable.

Test had to be adjusted for the changes, and also for the fact that
process() for hunk now uses pre-/post-image contents, if available,
therefore lexing results should be the same.

Added test for the case where hunk does not have access to
pre-/post-image, and where we continue to expect lexing results to be
different (at least in that specific case of the patch, files, and
changes used for those tests).