annotate.py: Add `repo` parameter to Bug.from_patchset()
If the provided `repo` is the Git repository with the commit identified by
`patch_id` (assuming that changes in `patch_set` are changes in the
`patch_id` commit, e.g. from repo.unidiff(patch_id)), then it will be
used to provide pre-image and post-image contents for changed files in
`patch_id` changeset.

Currently added tests only check that new code corectly handles corner
cases: when repo is valid but not the one `patch_id` commit came from,
and when repo is invalid (is not Git repository).

TODO: check that the main part of added code functions correctly.