annotate.py: Add optional `repo` parameter to BugDataset contructor
This parameter is now provided in the BugDataset.from_repo() class
method (alternate contructor).  It is saved, explicitely, in the
_git_repo field of BugDataset.

Add boolean `use_repo` parameter to the BugDataset.get_bug() method;
if true (the default), pass the _git_repo to Bug.from_patchset()
for use.  This would make it easy to add a command line parameter that
controls whether to use the repository for pre-/post-images.

No specific tests for this behavior yet.