utils.git: Add `prev` field to ChangeSet class
It is set by unidiff() method, in order to be able to distinguish
when the diff might have come from a commit, as opposed to diff
between two commits further away.

The default for `prev` is None; log_p() method does not set `prev`.

While at it, add docstring for ChangeSet primary constructor.

No test for the new feature.