gather_data.py: Add --purpose-to-annotation option to 'timeline'
This option is repeatable, and can be passed multiple times.  The
values expected to be in the "<file purpose>:<line type>" form.

It can be used to make every line of files with selected purposes
to have all selected type.  For example, we can make every line of
files with 'test' purpose to have 'test' type by passing the
`--purpose-to-annotation=test:test`.

Click was explicitly added to requirements, but it was already
transitively required by typer / typer-slim.