annotate.py: Add 'spread_inner' to compute_sizes_and_spread
This is the only part of patch spread that can be computed by single
hunk.  Full compute of spread requires computing number of lines between
hunks, or between last and first group in a pair of adjacent hunks
in a patched file.

Update TODOs in the docstring.

.
While at it, remove print() leftover from debugging, which was on
a branch which should never happen in a proper unified diff.

https://www.gnu.org/software/diffutils/manual/html_node/Detailed-Unified.html