annotate.py: Add tests, fix counting 'inner_sep' for hunk
Newly added test found a bug in the code computing 'inner_sep' in
AnnotatedHunk.compute_sizes_and_spreads() method, namely that the
count of context lines was not reset after the end of a group of
context lines (that might turn out to be separating groups of changes).

This was visible only when examining its behavior for a patched file
with hunk that had more than 2 groups of changed lines.

Also adds some commented out debugging 'print's.