annotate.py: Compute 'groups_spread' that includes inter-hunk distances
The 'groups_spread', which is one of metrics for patch spreading,
sum of "number of lines interleaving chunks [(groups)] in a patch"
over all changed files.  It is sum of inner hunk spread and inter-hunk
distance.

Update docstrings to mark that this feature is no longer on TODO list,
and is no longer denoted TBD (To Be Done).

Adds two new checks for single-file 'groups_spread':
- for the case where in neighbouring hunks, one hunk ends in the same
  line type as the next hunk begins (e.g. '-' and '-'), and
- for the case where one hunk ends in different line type that
  the next hunk begins (e.g. '+' and '-').
Those checks were added to the test_annotate.py test case, to the test
called now test_misc_patchsets_sizes_and_spreads().