annotate.py: Properly configure logging, add logging to file
When running the 'diff-annotate' script that was generated by the
install process, in annotate.py the __name__ variable is not '__main__'.
Therefore to set up logging to write to a file, one needs to get
root logger, and manipulate root logger in a common() function, which
is common entry point for all subcommand of this script.

Make printing information about logging level more robust; do not rely
on log level not changing (remove duplication - two subsequent lines).

NOTE however that log output is still printed to the console, but
is also written to the 'diff-annotate.log' log file.  Perhaps using
force=True in .basicConfig() would help...