test_cli.py: Make it easier to find the true source of errors
Now all CLI tests print name of exception and traceback if the command
failed because of an uncaught exception.

This makes it easier to find the ultimate source of the error; the
exception alone (with its message) shown by failing

   assert result.exit_code == 0

is often not enough.

TODO: refactor to avoid code duplication.