01-gather_data-explore.ipynb: Add xticks to heatmap, show other author
Comparing "timeline" data for two different authors using the sns.heatmap()
is complicated by the fact that this plot is more bar-like than scatter-like,
and it does not respect the ax.set_xlim().  Therefore, to align x axis ticks
when different authors starts and ends at different time, one has to extend
the dataframe adding zeros for the time period where there were no changes.

This technique of extending dateframe with zeros to match start and end dates
only started to be tested...