dvc.yaml: Add "annotate" stage
This stage runs `diff-annotate from-repo ...` on repositories cloned
in the preceding "clone" stage.  This stage also makes use of 'foreach'
stages, and of 'vars'.

Do not use external outputs, DVC 3.x does not like it (and it gets
confused to the point of `dvc stage` failing, with a strange error message).