Add notebooks/panel/00-panel-simple-Matplotlib.ipynb
This is a very simple example of a semi-interactive plot, with widgets
and display provided by the Panel, and drawing done with matplotlib
(using the Matplotlib pane).

Based on first example in  https://panel.holoviz.org/reference/panes/Matplotlib.html

At the time of being, it works correctly only in 'Preview with Panel',
but does show almost empty page when run via `panel serve ...`.