Add notebooks/panel/01-timeline.ipynb notebook
This notebook is meant to be run with Panel, either via 'Preview with Panel'
action from JupyterLab toolbar, or with `panel serve ...`.

When previewed / served with Panel, it shows two plots: counts of '+'
and '-' lines, resampled, and aggregated, and count of commits, resampled.
You can change the following parameters with widgets on left hand side:

- the repository to be analyzed, inactive and unused
- the author to be analyzed (one of 2 available)
- resampling period, e.g. monthly, or weekly
- aggregation function over sampling period, e.g. 'sum' over month

While at it, add to notebooks/.gitignore file pattern for output files
created by Panel.