panel/01-timeline.ipynb: Add heatmap plot - line type vs time
The new plot does not yet include proper formatting of labels, and
as it turns out it doesn't look well for resampling rates other than 'ME'
(styling was specifically adjusted for that specific sampling rate).

Also adds the beginnings of bihistogram of +/- lines, plotting histogram
of number of added and removed lines per commit, or per resampling period.
Some plots were generated, but they are not yet added to the Panel app:
the label and/or titles lack some neccessary information, like info about
the aggregation function used.