panel/01-timeline.ipynb: Add histogram of -/+ lines count per commit
Note that it is taken directly from the original data, not from
the dataframe after resampling.  Therefore the plot does not depend
on the resampling rate.

Widgets for controlling the histogram parameters (bin width, max value)
are nicely grouped together, and throttled.

TODO: use the same chain of bound functions technique that histogram
plot uses also for other plots in the app.