README.md: Add GitHub Insights in the "Related projects" section
This part describes the web tool in more detail than other tools, with
the goal of creating a web page in the demo web app for diff-annotator
that provides similar information to GitHub, but enhanced.

TODO: consider moving the whole "Related projects" section to a separate
file, perhaps in the planned 'docs/' folder.