panel/01-timeline.ipynb: Add histogram of -/+ counts per resample period
This histogram was started in one of prior commmits, it was just finalized,
bound with widgets, and added to the dashboard in this commit.  One of
the issues that needed to be solved was showing the aggregation function
used/histogrammed.

While at it, use Unicode minus '−' instead of ASCII dash '-' in legend,
as it is the same width as the Unicode/ASCII plus '+' sign, which leads
to nicer layout.