panel/01-timeline.ipynb: Add -/+ [%] columns and make them available
In addition to being able to view timeline of aggregate of selected
-/+ lines (for example: -/+:type.code) over selected resample period,
it is now possible to view aggregate of percentage of selected -/+ lines.
For example, one can now view aggregate of -/:type.code [%], which
is simply -/+:type.code / -/+:count.

The y-range for -/+ [%] columns always span from 0.0 to 1.0 (that is,
from 0% to 100%).  There is no autoscaling for this types of columns.

(Re)computing of -/+ [%] columns is done with add_pm_count_perc(),
added in the new "Percentages" subsection.  The plot_pm_col(), added
earlier, now has a special y-range handling for -/+ [%] columns.

The column_base_widget now includes -/+ [%] columns as choice
(possibly disabled, i.e. visible but unselectable).

.
Unfortunately, it turns out that there is some problem with the
way reactive components and dependencies are set up, in that resampling
is unnecessary re-done too many times for a single change of
the column (field) to display.  This in turns meas that -/+ [%] columns
also need to be recalculated.  The plot_pm_col() now protects against
this issue, instead of relying on set-up dependencies to do this job.

In the definition of bound_plot_pm_col, `resampled_df=bound_add_pm_count_perc`
should be used in place of `resampled_df=bound_resample_timeline`
(it is now commented out).

TODO: debug why resample_timeline() and add_pm_count_perc() are
called multiple times for a single change of widgets.