panel/01-timeline.ipynb: Make more robust wrt. `panel serve` workdir
Try to find requested JSON file somewhere in the current directory
and limited list of parent directories.  This makes the notebook
robust with respect to the working directory of `panel serve` command,
that is the directory `panel serve` was invoked from: top directory
of the project, 'notebooks/` directory, or 'notebooks/panel/' subdirectory.

Previously the notebook assumed that `panel serve` was invoked from
the directory with the notebook, namely 'notebooks/panel/'.