panel/01-timeline.ipynb: Debbuging with JSONEdit widget in view mode
Adds JSONEdit widget in a view mode inside a Card, separated with Saparator.
Its presence is controlled with the DEBUG global variable, currently set
to True.  Required adding 'jsoneditor' extension to Panel.

NOTE: This is to be a commonn code for all notebooks and scripts in
the 'notebooks/panel/' directory.