panel/01-timeline.ipynb: Debbuging with Tabulator showing tf_timeline_df
Add a Tabulator widget showing the tf_timeline_df DataFrame, with
the data about single repository ('dataset') in it.  This tabular data
is extracted from the JSON file that the widget added in previous commit
shows.

Alternative would be to use the (deprecated) DataFrame widget.

NOTE that the Tabulator widget has its width set to a specific value, instead
of configuring it to span the whole width of the container (of the main pane),
because it did't work - no scrolbar, no limiting to viewport width.

https://panel.holoviz.org/reference/widgets/Tabulator.html
https://tabulator.info/