panel/01-timeline.ipynb: Debbuging with Perspective showing resampled df
The Perspective pane[1] provides a powerful visualization component for large,
real-time datasets built on the Perspective Project[2].

[1]: https://panel.holoviz.org/reference/panes/Perspective.html
[2]: https://perspective.finos.org/

Add a Perspective pane at the bottom of the main part of the page,
which displays resampled data (using `bound_resample_timeline` bound
function).  This allows for easier debugging of various issues with
the mentioned DataFrame.