panel/01-timeline.ipynb: Debbuging with logger and Debugger widget
Setting up the pn.widgets.Debugger widget was tested with the
panel/00-panel-Debugger-demo.ipynb Jupyter notebook.

Logging (logger.debug) was added to functions that load data, create
dataframes, or alter dataframes.

Examining the logger output have shown the following problems:
- get_timeline_data is cached, so (in most cases) is not seen in logs
- every change of the author, or of the sampling rate, leads to
  11 (!) invocations of resample_timeline
- switching toa -/+ [%] column (for some of plots) leads to
  2 invocations of add_pm_count_perc, and 2 of add_pm_count_perc
  for every switch, and not only for the first one
- even toggling autoscaling (turning it on or off) leads to
  2 invocations of add_pm_count_perc, and 2 of add_pm_count_perc

This needs to be solved, either with pn.rx, or with pn.cache (or both).