annotate.py: Add optional `repo` field to AnnotatedPatchSet
You can either provide it as an argument in the constructor, or use
the add_repo() method (which can be used with a builder pattern).

This would help AnnotatedPatchSet absorb all functionality, and make
the Bug class just add on saving annotations to a file etc.  This would
help remove currently occuring code duplication.