annotate.py: Add missing functionality to AnnotatedPatchSet.process()
Now the .process() method in AnnotatedPatchSet class will try to read
the pre-image and post-image contents of file from the Git repository,
and use it when processing each patched file.  This functionality
is present in Bug.from_patchset() class method, but was missing in
AnnotatedPatchSet.process().

This change required adding the `repo` field to AnnotatedPatchSet
(which was added in previous commit), if the constraint is to not
change the signature of the .process() method.

Note that as a side effect some work was extracted outside
try...except... block.