annotate.py: Separate different data in AnnotatedPatchSet.process() output
The commit metadata was already put whole as a dict under
'commit_medatata' key in the .process() method output.

Do the same for information about changes - putting them under 'changes',
and, if requested, patch size and spread metrics - putting them under
'diff_metadata'.

The change is so simple because of previous commits in the series.

WARNING: this *breaks* most if not all of gather_data.py script because
it is schema change!!! This is not fixed yet!

While at it, add beginnings of tests for correctness of computing
patch size and spread metrics - currently marked as expected failure,
because the bug is not yet fixed.