test_annotate.py: Add test for -/+ lines - 'changes' vs 'diff_metadata'
Check if the sum of number of '-' and '+' lines in the diff (extracted from
'changes' in annotated data) matches 'n_rem' + 2*'n_mod' + 'n_add'
metrics (extracted from 'diff_metadata' in annotated data), at least for
the simple case of 3 latest commits in githubtraining/hellogitworld repository.

The equation

  '-' + '+' == 'n_rem' + 2*'n_mod' + 'n_add'

follows from the fact that each modified line is represented in unified
diff as removal ('-') of the old version, and addition ('+') of the new
version of the line.