test_annotate.py: Test 'changes' vs 'diff_metadata' for multi-file diff
The exact patch file (*.diff) to be tested was selected based on
examination of mismatch between -/+:count and n_rem, n_mod, n_add
values found in notebooks/panel/01-timeline.ipynb, section "Patch size
and spread, resampled".

The fact that no discrepancies were found in the test (added to the
tests/test_annotate.py::test_misc_patchsets_sizes_and_spreads test)
most likely measn that the bug was not in annotate.py - thus removal
of the task from TODO.md.

(It might be the case that unification of the code computing of patch
size and spread metrics in .process() fixed the issue instead, or that
if fixed the issue for the gather_data.py too.)