gather_data.py: Introduce `data_format` parameter to callbacks
These are callbacks passed as parameter to .gather_data and friend,
namely:
- PurposeCounterResults.create()
- map_diff_to_purpose_dict()
- map_diff_to_lines_stats()
- map_diff_to_timeline()

The `data_format` stays at its default value, which is the same as
default value in _maybe_changes() and _maybe_diff_data().  This value
is passed down to _maybe_changes() and _maybe_diff_data() - which means
that the code should not change it's behavior.

This change is a refactoring.

NOTE: This commit prepares for the next step, where format version will
be guessed from the name of input files, overriding the default value
for `data_format` parameter.