test_annotate.py: Extract count_pm_lines() and move it to conftest.py
Extracting this function reduces code duplication (one of which was
slightly wrong in a way that didn't make the test it was in to fail:
total_p and total_m were switched around in that copy of the code).

Moving it conftest.py means that it would be able to be used by other
tests, like new test in test_cli.py that will be added in next commit.