tests: Fix bug in example_line_callback_func.py
Line is empty if it is somposed of a single token, and the text fragment
of this token (which is 3rd element of the tuple = index 2) is "\n".

This fixes tests failing because of bad interaction between tests,
where test checking that providing line callback works (not hitting the
bug), but some subsequent test somehow inherits the callback - and the
bug is hit for this test.