panel/01-timeline.ipynb: Switch from pn.bind to pn.rx
The Panel documentation recomments using reactive components instead
of bounding functions, especially instead of binding to bound functions.
The latter can help you to quickly explore your data, but it can be
(and is) inefficient as the results are calculated from scratch for each
call.

This commit brings significant performance improvements with regard
to the number of re-calculations.

Before (see 3704c90b02):
- every change of the author, or of the sampling rate, leads to
  11 (!) invocations of resample_timeline
- switching to a -/+ [%] column (for some of plots) leads to
  2 invocations of add_pm_count_perc, and 2 of add_pm_count_perc
  for every switch, and not only for the first one
- even toggling autoscaling (turning it on or off) leads to
  2 invocations of add_pm_count_perc, and 2 of add_pm_count_perc

After this change:
- every change of the author leads to 2 invocations of resample_timeline
  (one for selected resample rate, one for 'ME' resample rate)
  and 1 invocation of add_pm_count_perc
- every change to resample rate leads to 1 invocation of resample_timeline
  and 1 invocation of add_pm_count_perc
- switching to -/+ [%] column (available for come of plots),
  or toggling autoscaling, do not lead to any recalculations

Another solution would be to use pn.cache, perhaps limited to session
for the ease of development.  However, some places used pn.rx for
convenience already in addition to using pn.bind.  Adding @pn.cache
is planned for the future commit - it speeds up the application,
but might interfere with the development process (serving stale data
from the cache after code changes).

This commit switches from using pn.bind, to using pn.rx for all plots.

https://panel.holoviz.org/tutorials/basic/pn_bind.html
https://panel.holoviz.org/tutorials/basic/pn_rx.html
https://panel.holoviz.org/tutorials/intermediate/parameters.html
https://param.holoviz.org/user_guide/Reactive_Expressions.html

.
NOTE: There is some strange problem that `panel serve` detects with
the notebook, but it is not something critical:

  python-diff-annotator/notebooks/panel/01-timeline.ipynb:1641:
    SyntaxWarning: invalid escape sequence '\['
    "metadata": {

Maybe it could be fixed by hand editing *.ipynb file...